refactor(components): refactor PopupMsg component
